+++ account/nuls-account/src/main/java/io/nuls/account/service/AccountService.java
@@ -155,23 +155,6 @@ public interface AccountService {
     List<String> getAllPrivateKey(short chainId, String password);
 
     /**
-     * 从keyStore导入账户(密码用来验证keystore)
-     * 1.从keyStore获取明文私钥(如果没有明文私钥,则通过密码从keyStore中的encryptedPrivateKey解出来)
-     * 2.通过keyStore创建新账户,加密账户
-     * 3.从数据库搜索此账户的别名,没有搜到则不设置(别名不从keyStore中获取,因为可能被更改)
-     * 4.保存账户
-     * import an account form account key store.
-     *
-     * @param keyStore  the keyStore of the account.
-     * @param chainId
-     * @param password  the password of account
-     * @param overwrite
-     * @return the result of the operation.
-     * @throws NulsException
-     */
-    Account importAccountFormKeyStore(AccountKeyStore keyStore, short chainId, String password, boolean overwrite) throws NulsException;
-
-    /**
      * 根据私钥和密码导入账户
      * import an account from plant private key and encrypt the account.
      *
@@ -184,13 +167,4 @@ public interface AccountService {
      */
     Account importAccount(short chainId, String prikey, String password, boolean overwrite) throws NulsException;
 
-    /**
-     * 导出账户到keyStore
-     * export an account to an account key store.
-     *
-     * @param address  the address of the account.
-     * @param password the password of the account key store.
-     * @return the account key store object.
-     */
-    AccountKeyStore exportAccountToKeyStore(short chainId, String address, String password);
 }
