+++ account/nuls-account/src/main/java/io/nuls/account/service/impl/AccountKeyStoreServiceImpl.java
@@ -1,0 +1,188 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2017-2018 nuls.io
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ */
+
+package io.nuls.account.service.impl;
+
+import io.nuls.account.constant.AccountErrorCode;
+import io.nuls.account.model.bo.Account;
+import io.nuls.account.model.bo.AccountKeyStore;
+import io.nuls.account.model.po.AccountPo;
+import io.nuls.account.model.po.AliasPo;
+import io.nuls.account.service.AccountCacheService;
+import io.nuls.account.service.AccountKeyStoreService;
+import io.nuls.account.service.AccountService;
+import io.nuls.account.storage.AccountStorageService;
+import io.nuls.account.storage.AliasStorageService;
+import io.nuls.account.util.AccountTool;
+import io.nuls.base.basic.AddressTool;
+import io.nuls.base.data.Address;
+import io.nuls.tools.basic.InitializingBean;
+import io.nuls.tools.core.annotation.Autowired;
+import io.nuls.tools.core.annotation.Service;
+import io.nuls.tools.crypto.AESEncrypt;
+import io.nuls.tools.crypto.ECKey;
+import io.nuls.tools.crypto.HexUtil;
+import io.nuls.tools.data.FormatValidUtils;
+import io.nuls.tools.data.StringUtils;
+import io.nuls.tools.exception.CryptoException;
+import io.nuls.tools.exception.NulsException;
+import io.nuls.tools.exception.NulsRuntimeException;
+import io.nuls.tools.log.Log;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
+
+/**
+ * @author: qinyifeng
+ */
+@Service
+public class AccountKeyStoreServiceImpl implements AccountKeyStoreService {
+
+    @Autowired
+    private AccountService accountService;
+
+    @Autowired
+    private AccountStorageService accountStorageService;
+
+    @Autowired
+    private AliasStorageService aliasStorageService;
+
+    private AccountCacheService accountCacheService = AccountCacheService.getInstance();
+
+    @Override
+    public Account importAccountFormKeyStore(AccountKeyStore keyStore, short chainId, String password, boolean overwrite) throws NulsException {
+        //check params
+        if (null == keyStore || StringUtils.isBlank(keyStore.getAddress())) {
+            throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
+        }
+        if (!AddressTool.validAddress(chainId, keyStore.getAddress())) {
+            throw new NulsRuntimeException(AccountErrorCode.ADDRESS_ERROR);
+        }
+        if (StringUtils.isNotBlank(password) && !FormatValidUtils.validPassword(password)) {
+            throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
+        }
+        //not allowed to cover
+        if (!overwrite) {
+            Address address = AccountTool.newAddress(chainId, keyStore.getAddress());
+            //Query account already exists
+            Account account = accountService.getAccount(chainId, address.getBase58());
+            if (null != account) {
+                throw new NulsRuntimeException(AccountErrorCode.ACCOUNT_EXIST);
+            }
+        }
+        Account account;
+        byte[] priKey;
+        //if the private key is not encrypted, it is not empty
+        if (null != keyStore.getPrikey() && keyStore.getPrikey().length > 0) {
+            if (!ECKey.isValidPrivteHex(HexUtil.encode(keyStore.getPrikey()))) {
+                throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
+            }
+            //create account by private key
+            priKey = keyStore.getPrikey();
+            account = AccountTool.createAccount(chainId, HexUtil.encode(priKey));
+            //如果私钥生成的地址和keystore的地址不相符，说明私钥错误
+            //if the address generated by the private key does not match the address of the keystore, the private key error
+            if (!account.getAddress().getBase58().equals(keyStore.getAddress())) {
+                throw new NulsRuntimeException(AccountErrorCode.PRIVATE_KEY_WRONG);
+            }
+        } else if (null == keyStore.getPrikey() && null != keyStore.getEncryptedPrivateKey()) {
+            //加密私钥不为空,验证密码是否正确
+            //encrypting private key is not empty,verify that the password is correct
+            if (!FormatValidUtils.validPassword(password)) {
+                throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
+            }
+            try {
+                //create account by private key
+                priKey = AESEncrypt.decrypt(HexUtil.decode(keyStore.getEncryptedPrivateKey()), password);
+                account = AccountTool.createAccount(chainId, HexUtil.encode(priKey));
+            } catch (CryptoException e) {
+                throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
+            }
+            //如果私钥生成的地址和keystore的地址不相符，说明私钥错误
+            //if the address generated by the private key does not match the address of the keystore, the private key error
+            if (!account.getAddress().getBase58().equals(keyStore.getAddress())) {
+                throw new NulsRuntimeException(AccountErrorCode.PRIVATE_KEY_WRONG);
+            }
+        } else {
+            throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
+        }
+        //Query account already exists
+        Account acc = accountService.getAccount(chainId, account.getAddress().getBase58());
+        if (null == acc) {
+            //查询全网该链所有别名对比地址符合就设置
+            //query the whole network. All the aliases of the chain match the addresses
+            account.setAlias(accountService.getAliasByAddress(chainId, account.getAddress().getBase58()));
+        } else {
+            //if the local account already exists
+            account.setAlias(acc.getAlias());
+        }
+
+        //encrypting account private key
+        if (FormatValidUtils.validPassword(password)) {
+            account.encrypt(password);
+        }
+        //save account to db
+        accountStorageService.saveAccount(new AccountPo(account));
+        //put the account in local cache
+        accountCacheService.localAccountMaps.put(account.getAddress().getBase58(), account);
+        return account;
+    }
+
+    @Override
+    public AccountKeyStore exportAccountToKeyStore(short chainId, String address, String password) {
+        //check params
+        if (!AddressTool.validAddress(chainId, address)) {
+            throw new NulsRuntimeException(AccountErrorCode.ADDRESS_ERROR);
+        }
+        //check the account is exist
+        Account account = accountService.getAccount(chainId, address);
+        if (null == account) {
+            throw new NulsRuntimeException(AccountErrorCode.ACCOUNT_NOT_EXIST);
+        }
+        AccountKeyStore accountKeyStore = new AccountKeyStore();
+        //如果账户加密,就验证密码
+        //if the account is encrypted, verify the password
+        if (account.isEncrypted() && !account.validatePassword(password)) {
+            throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
+        }
+        //如果账户加密,都不导出明文私钥
+        //if the account is encrypted , the plaintext private key is not exported
+        if (account.isEncrypted()) {
+            accountKeyStore.setEncryptedPrivateKey(HexUtil.encode(account.getEncryptedPriKey()));
+        } else {
+            accountKeyStore.setPrikey(account.getPriKey());
+        }
+        accountKeyStore.setAddress(account.getAddress().toString());
+        accountKeyStore.setPubKey(account.getPubKey());
+        return accountKeyStore;
+    }
+}
