+++ account/nuls-account/src/main/java/io/nuls/account/service/impl/AccountServiceImpl.java
@@ -453,111 +453,4 @@ public class AccountServiceImpl implements AccountService, InitializingBean {
         return account;
     }
 
-    @Override
-    public Account importAccountFormKeyStore(AccountKeyStore keyStore, short chainId, String password, boolean overwrite) throws NulsException {
-        //check params
-        if (null == keyStore || StringUtils.isBlank(keyStore.getAddress())) {
-            throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
-        }
-        if (!AddressTool.validAddress(chainId, keyStore.getAddress())) {
-            throw new NulsRuntimeException(AccountErrorCode.ADDRESS_ERROR);
-        }
-        if (StringUtils.isNotBlank(password) && !FormatValidUtils.validPassword(password)) {
-            throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
-        }
-        //not allowed to cover
-        if (!overwrite) {
-            Address address = AccountTool.newAddress(chainId, keyStore.getAddress());
-            //Query account already exists
-            Account account = this.getAccountByAddress(chainId, address.getBase58());
-            if (null != account) {
-                throw new NulsRuntimeException(AccountErrorCode.ACCOUNT_EXIST);
-            }
-        }
-        Account account;
-        byte[] priKey;
-        //if the private key is not encrypted, it is not empty
-        if (null != keyStore.getPrikey() && keyStore.getPrikey().length > 0) {
-            if (!ECKey.isValidPrivteHex(HexUtil.encode(keyStore.getPrikey()))) {
-                throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
-            }
-            //create account by private key
-            priKey = keyStore.getPrikey();
-            account = AccountTool.createAccount(chainId, HexUtil.encode(priKey));
-            //如果私钥生成的地址和keystore的地址不相符，说明私钥错误
-            //if the address generated by the private key does not match the address of the keystore, the private key error
-            if (!account.getAddress().getBase58().equals(keyStore.getAddress())) {
-                throw new NulsRuntimeException(AccountErrorCode.PRIVATE_KEY_WRONG);
-            }
-        } else if (null == keyStore.getPrikey() && null != keyStore.getEncryptedPrivateKey()) {
-            //加密私钥不为空,验证密码是否正确
-            //encrypting private key is not empty,verify that the password is correct
-            if (!FormatValidUtils.validPassword(password)) {
-                throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
-            }
-            try {
-                //create account by private key
-                priKey = AESEncrypt.decrypt(HexUtil.decode(keyStore.getEncryptedPrivateKey()), password);
-                account = AccountTool.createAccount(chainId, HexUtil.encode(priKey));
-            } catch (CryptoException e) {
-                throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
-            }
-            //如果私钥生成的地址和keystore的地址不相符，说明私钥错误
-            //if the address generated by the private key does not match the address of the keystore, the private key error
-            if (!account.getAddress().getBase58().equals(keyStore.getAddress())) {
-                throw new NulsRuntimeException(AccountErrorCode.PRIVATE_KEY_WRONG);
-            }
-        } else {
-            throw new NulsRuntimeException(AccountErrorCode.PARAMETER_ERROR);
-        }
-        //Query account already exists
-        Account acc = getAccountByAddress(chainId, account.getAddress().getBase58());
-        if (null == acc) {
-            //查询全网该链所有别名对比地址符合就设置
-            //query the whole network. All the aliases of the chain match the addresses
-            account.setAlias(this.getAliasByAddress(chainId, account.getAddress().getBase58()));
-        } else {
-            //if the local account already exists
-            account.setAlias(acc.getAlias());
-        }
-
-        //encrypting account private key
-        if (FormatValidUtils.validPassword(password)) {
-            account.encrypt(password);
-        }
-        //save account to db
-        accountStorageService.saveAccount(new AccountPo(account));
-        //put the account in local cache
-        accountCacheService.localAccountMaps.put(account.getAddress().getBase58(), account);
-        return account;
-    }
-
-    @Override
-    public AccountKeyStore exportAccountToKeyStore(short chainId, String address, String password) {
-        //check params
-        if (!AddressTool.validAddress(chainId, address)) {
-            throw new NulsRuntimeException(AccountErrorCode.ADDRESS_ERROR);
-        }
-        //check the account is exist
-        Account account = this.getAccountByAddress(chainId, address);
-        if (null == account) {
-            throw new NulsRuntimeException(AccountErrorCode.ACCOUNT_NOT_EXIST);
-        }
-        AccountKeyStore accountKeyStore = new AccountKeyStore();
-        //如果账户加密,就验证密码
-        //if the account is encrypted, verify the password
-        if (account.isEncrypted() && !account.validatePassword(password)) {
-            throw new NulsRuntimeException(AccountErrorCode.PASSWORD_IS_WRONG);
-        }
-        //如果账户加密,都不导出明文私钥
-        //if the account is encrypted , the plaintext private key is not exported
-        if (account.isEncrypted()) {
-            accountKeyStore.setEncryptedPrivateKey(HexUtil.encode(account.getEncryptedPriKey()));
-        } else {
-            accountKeyStore.setPrikey(account.getPriKey());
-        }
-        accountKeyStore.setAddress(account.getAddress().toString());
-        accountKeyStore.setPubKey(account.getPubKey());
-        return accountKeyStore;
-    }
 }
